maya custom dynamics outliner
I do hair and cloth. Sometimes because the Maya scene is too complicated, Maya has many objects in the default Outliner, but sometimes I only want to list the dynamics nodes of hairSystem, nCloth, and nucleus, so I want to have a custom dynamics Outliner. The dynamics Outliner only stores the dynamics node, which is very convenient for me to choose, but there is a problem. For example, when I select the dynamics node in the dynamics Outliner, when I select an object other than the dynamics Outliner, Maya will select the object and cancel it immediately select

Please g'w...